// payment-service/src/utils/broker/type.ts
import { MessageType, PaymentEvent, TOPIC_TYPE } from "~/types/subscription"

export interface PublishType {
  headers: Record<string, any>
  topic: TOPIC_TYPE
  event: PaymentEvent
  message: Record<string, any>
}

export type MessageHandler = (input: MessageType) => void

export type MessageBrokerType = {
  // Producer
  connectProducer: <T>() => Promise<T>
  disconnectProducer: () => Promise<void>
  publish: (data: PublishType) => Promise<boolean>
}